requirements.txt
httpx~=0.28.1
beautifulsoup4~=4.12.3
fastapi~=0.115.12
uvicorn~=0.22.0
dotenv~=0.9.9
python-dotenv~=1.1.1
pillow~=11.3.0